+++ src/world/builders/index.ts
@@ -1,0 +1,7 @@
+export { FloorBuilder } from "./FloorBuilder";
+export type { FloorConfig } from "./FloorBuilder";
+
+export { WallBuilder } from "./WallBuilder";
+export type { WallBuilderConfig } from "./WallBuilder";
+
+export { DoorOpeningBuilder } from "./DoorOpeningBuilder";